Add Ingredient and MarketOrder models and implements ingredients buy
New Ingredient and MarketOrder models are added. Functionality for purchasing ingredients from the market is implemented. Also, Sanctum for API authentication and related configuration is added. The IngredientsController is updated to handle storing ingredient and buying process. Minor changes in database migrations and seeder files are done to streamline the system.