Add 'Resumen' page and update related components and dependencies
Implemented new 'Resumen' page providing statistical summary about data. This involved an update in the Navbar component for a new navigation link. It also required the creation of new related css files and components, and addition of new chart-related dependencies in package.json. A new 'Resumen' API route was created to fetch the necessary data.